requirements.txt
streamlit
pdfplumber
python-docx
scikit-learn
rapidfuzz
